+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## main
+
+- Added the `AppStore.Token` for token generation.
+
 ## v0.1.0
 
 Initial release